+++ src/dynd/dtypes/convert_dtype.cpp
@@ -71,7 +71,7 @@ dtype dynd::convert_dtype::apply_linear_index(int nindices, const irange *indice
     }
 }
 
-void dynd::convert_dtype::get_shape(int i, intptr_t *out_shape) const
+void dynd::convert_dtype::get_shape(size_t i, intptr_t *out_shape) const
 {
     if (!m_value_dtype.is_builtin()) {
         m_value_dtype.extended()->get_shape(i, out_shape);
